refactor: separate domain and view package
