+++ src/main/kotlin/racingcar/Controller.kt
@@ -1,5 +1,7 @@
 package racingcar
 
+import racingcar.domain.RacingManager
+
 fun main() {
     RacingManager().run()
 }
